move all payments-related stuff to a Payments module
