+++ lib/banchan/commissions/commissions.ex
@@ -23,6 +23,7 @@ defmodule Banchan.Commissions do
 
   alias Banchan.Offerings
   alias Banchan.Offerings.{Offering, OfferingOption}
+  alias Banchan.Payments.Invoice
   alias Banchan.Repo
   alias Banchan.Studios
   alias Banchan.Studios.Studio
@@ -361,12 +362,6 @@ defmodule Banchan.Commissions do
   end
 
   @doc """
-  True if a given invoice has been paid. Expects an already-loaded invoice
-  with the latest data.
-  """
-  def invoice_paid?(%Invoice{status: status}), do: status == :succeeded
-
-  @doc """
   Calculates the total currently deposited amount for a commission.
   """
   def deposited_amount(
@@ -698,51 +693,10 @@ defmodule Banchan.Commissions do
     ret
   end
 
-  @doc """
-  Creates a new invoice.
-  """
-  def invoice(actor, commission, current_user_member?, drafts, event_data)
-
-  def invoice(actor, commission, false, drafts, event_data) do
-    if :admin in actor.roles || :mod in actor.roles do
-      invoice(actor, commission, true, drafts, event_data)
-    else
-      {:error, :unauthorized}
-    end
-  end
-
-  def invoice(%User{} = actor, %Commission{} = commission, true, drafts, event_data) do
-    {:ok, ret} =
-      Repo.transaction(fn ->
-        actor = Repo.reload(actor)
-
-        if Studios.is_user_in_studio?(actor, %Studio{id: commission.studio_id}) ||
-             :admin in actor.roles ||
-             :mod in actor.roles do
-          with {:ok, event} <-
-                 create_event(:comment, actor, commission, true, drafts, event_data),
-               %Invoice{} = invoice <- %Invoice{
-                 client_id: commission.client_id,
-                 commission: commission,
-                 event: event
-               },
-               %Ecto.Changeset{} = changeset <- Invoice.creation_changeset(invoice, event_data),
-               {:ok, invoice} <- Repo.insert(changeset) do
-            send_event_update!(event.id, actor)
-            {:ok, invoice}
-          end
-        else
-          {:error, :unauthorized}
-        end
-      end)
-
-    ret
-  end
-
   ## Edit/Update
 
   # credo:disable-for-next-line Credo.Check.Refactor.CyclomaticComplexity
-  defp check_actor_edit_access(%User{} = actor, %Commission{} = commission) do
+  def check_actor_edit_access(%User{} = actor, %Commission{} = commission) do
     actor = actor |> Repo.reload() |> Repo.preload(:disable_info, force: true)
 
     cond do
@@ -862,52 +816,6 @@ defmodule Banchan.Commissions do
               {:ok, {commission, [event]}}
             end
           end
-        end
-      end)
-
-    ret
-  end
-
-  @doc """
-  Releases a payment.
-  """
-  def release_payment(%User{} = actor, %Commission{} = commission, %Invoice{} = invoice) do
-    {:ok, ret} =
-      Repo.transaction(fn ->
-        with {:ok, actor} <- check_actor_edit_access(actor, commission),
-             {1, _} <-
-               from(i in Invoice,
-                 join: c in assoc(i, :commission),
-                 where:
-                   i.id == ^invoice.id and c.client_id == ^actor.id and i.status == :succeeded
-               )
-               |> Repo.update_all(set: [status: :released]),
-             %Event{} = ev <-
-               from(e in Event,
-                 where: e.id == ^invoice.event_id,
-                 select: e,
-                 preload: [:actor, invoice: [], attachments: [:upload, :thumbnail, :preview]]
-               )
-               |> Repo.one() do
-          Notifications.invoice_released(commission, ev, actor)
-          Notifications.commission_event_updated(commission, ev, actor)
-
-          {:ok, ev}
-        end
-        |> case do
-          {0, _} ->
-            # Only succeeded invoices can be released.
-            {:error, :invalid_invoice_status}
-
-          {num, _} when is_integer(num) ->
-            # NB(zkat): This would point to a bug.
-            {:error, :updated_too_many_invoices}
-
-          {:ok, val} ->
-            {:ok, val}
-
-          {:error, err} ->
-            {:error, err}
         end
       end)
 
@@ -1202,443 +1110,6 @@ defmodule Banchan.Commissions do
     Event.text_changeset(event, attrs)
   end
 
-  defp send_event_update!(event_id, actor \\ nil) do
-    event =
-      from(e in Event,
-        where: e.id == ^event_id,
-        select: e,
-        preload: [:actor, invoice: [], attachments: [:upload, :thumbnail, :preview]]
-      )
-      |> Repo.one!()
-
-    commission =
-      from(c in Commission, where: c.id == ^event.commission_id)
-      |> Repo.one!()
-
-    Notifications.commission_event_updated(commission, event, actor)
-  end
-
-  @doc """
-  Processes a client payment
-  """
-  def process_payment(%User{id: user_id}, _, %Commission{client_id: client_id}, _, _)
-      when user_id != client_id do
-    {:error, :unauthorized}
-  end
-
-  def process_payment(
-        %User{} = actor,
-        %Event{invoice: %Invoice{amount: amount} = invoice} = event,
-        %Commission{} = commission,
-        uri,
-        tip
-      ) do
-    items = [
-      %{
-        name: "Commission Invoice Payment",
-        quantity: 1,
-        amount: amount.amount,
-        currency: String.downcase(to_string(amount.currency))
-      },
-      %{
-        name: "Extra Tip",
-        quantity: 1,
-        amount: tip.amount,
-        currency: String.downcase(to_string(tip.currency))
-      }
-    ]
-
-    with {stripe_id, platform_fee} <-
-           from(studio in Studio,
-             where: studio.id == ^commission.studio_id,
-             select: {studio.stripe_id, studio.platform_fee}
-           )
-           |> Repo.one(),
-         platform_fee <- Money.multiply(Money.add(amount, tip), platform_fee),
-         transfer_amt <- amount |> Money.add(tip) |> Money.subtract(platform_fee),
-         {:ok, session} <-
-           stripe_mod().create_session(%{
-             payment_method_types: ["card"],
-             mode: "payment",
-             cancel_url: uri,
-             success_url: uri,
-             line_items: items,
-             payment_intent_data: %{
-               transfer_data: %{
-                 amount: transfer_amt.amount,
-                 destination: stripe_id
-               }
-             }
-           }),
-         {:ok, _} <-
-           invoice
-           |> Invoice.submit_changeset(%{
-             tip: tip,
-             platform_fee: platform_fee,
-             stripe_session_id: session.id,
-             checkout_url: session.url,
-             status: :submitted
-           })
-           |> Repo.update() do
-      send_event_update!(event.id, actor)
-
-      {:ok, session.url}
-    end
-  end
-
-  @doc """
-  Webhook handler for when a payment has been successfully processed by Stripe.
-  """
-  def process_payment_succeeded!(session) do
-    {:ok, %{charges: %{data: [%{balance_transaction: txn_id, transfer: transfer}]}}} =
-      stripe_mod().retrieve_payment_intent(session.payment_intent, %{}, [])
-
-    {:ok, %{available_on: available_on, amount: amt, currency: curr}} =
-      stripe_mod().retrieve_balance_transaction(txn_id, [])
-
-    {:ok, transfer} = stripe_mod().retrieve_transfer(transfer)
-
-    total_charged = Money.new(amt, String.to_atom(String.upcase(curr)))
-    final_transfer_txn = transfer.destination_payment.balance_transaction
-
-    total_transferred =
-      Money.new(
-        final_transfer_txn.amount,
-        String.to_atom(String.upcase(final_transfer_txn.currency))
-      )
-
-    {:ok, available_on} = DateTime.from_unix(available_on)
-
-    {:ok, _} =
-      Repo.transaction(fn ->
-        {1, [invoice]} =
-          from(i in Invoice, where: i.stripe_session_id == ^session.id, select: i)
-          |> Repo.update_all(
-            set: [
-              status: :succeeded,
-              payout_available_on: available_on,
-              total_charged: total_charged,
-              total_transferred: total_transferred
-            ]
-          )
-
-        event = Repo.get!(Event, invoice.event_id)
-        client = Repo.reload!(%User{id: invoice.client_id})
-        commission = Repo.reload!(%Commission{id: event.commission_id})
-
-        create_event(
-          :payment_processed,
-          client,
-          commission,
-          true,
-          [],
-          %{
-            amount: invoice.amount |> Money.add(invoice.tip)
-          }
-        )
-
-        if client.email do
-          Notifications.send_receipt(invoice, client, commission)
-        end
-
-        send_event_update!(invoice.event_id)
-      end)
-
-    :ok
-  end
-
-  @doc """
-  Webhook handler for when a payment has been expired by a Studio member.
-  """
-  def process_payment_expired!(session) do
-    # NOTE: This will crash (intentionally) if we try to expire an
-    # already-succeeded payment. This should never happen, though, because it
-    # doesn't make sense that Stripe would tell us that a payment succeeded
-    # then tell us it's expired. So this crash is a good just-in-case.
-    {1, [invoice]} =
-      from(i in Invoice,
-        where: i.stripe_session_id == ^session.id and i.status != :succeeded,
-        select: i
-      )
-      |> Repo.update_all(set: [status: :expired])
-
-    send_event_update!(invoice.event_id)
-  end
-
-  @doc """
-  Expires a payment, preventing the client from being able to complete their
-  session, if it's still active.
-  """
-  def expire_payment(actor, invoice, current_user_member?)
-
-  def expire_payment(%User{} = actor, invoice, false) do
-    if :admin in actor.roles || :mod in actor.roles do
-      expire_payment(actor, invoice, true)
-    else
-      {:error, :unauthorized}
-    end
-  end
-
-  # credo:disable-for-next-line Credo.Check.Refactor.CyclomaticComplexity
-  def expire_payment(
-        %User{} = actor,
-        %Invoice{} = invoice,
-        true
-      ) do
-    {:ok, ret} =
-      Repo.transaction(fn ->
-        actor = actor |> Repo.reload()
-        invoice = invoice |> Repo.reload() |> Repo.preload(:commission)
-
-        if Studios.is_user_in_studio?(actor, %Studio{id: invoice.commission.studio_id}) ||
-             :admin in actor.roles || :mod in actor.roles do
-          case invoice do
-            %Invoice{id: id, stripe_session_id: nil, status: :pending} ->
-              # NOTE: This will crash (intentionally) if we try to expire an already-succeeded payment.
-              {1, [invoice]} =
-                from(i in Invoice, where: i.id == ^id and i.status != :succeeded, select: i)
-                |> Repo.update_all(set: [status: :expired])
-
-              send_event_update!(invoice.event_id)
-              {:ok, invoice}
-
-            %Invoice{stripe_session_id: session_id, status: :submitted} when session_id != nil ->
-              # NOTE: We don't manually expire the invoice in the database here. That's
-              # handled by process_payment_expired!/1 when the webhook fires.
-              with {:ok, _} <- stripe_mod().expire_payment(session_id) do
-                {:ok, invoice}
-              end
-
-            _ ->
-              {:error, :invalid_state}
-          end
-        else
-          {:error, :unauthorized}
-        end
-      end)
-
-    ret
-  end
-
-  @doc """
-  Refunds a payment that hasn't been released yet.
-  """
-  def refund_payment(actor, invoice, current_user_member?)
-
-  def refund_payment(actor, invoice, false) do
-    if :admin in actor.roles || :mod in actor.roles do
-      refund_payment(actor, invoice, true)
-    else
-      {:error, :unauthorized}
-    end
-  end
-
-  def refund_payment(%User{} = actor, %Invoice{} = invoice, _) do
-    {:ok, ret} =
-      Repo.transaction(fn ->
-        actor = actor |> Repo.reload()
-        invoice = invoice |> Repo.reload() |> Repo.preload(:commission)
-
-        if Studios.is_user_in_studio?(actor, %Studio{id: invoice.commission.studio_id}) ||
-             :admin in actor.roles || :mod in actor.roles do
-          if invoice.status != :succeeded do
-            Logger.error(%{
-              message: "Attempted to refund an invoice that wasn't :succeeded.",
-              invoice: invoice
-            })
-
-            {:error, :invoice_not_refundable}
-          else
-            process_refund_payment(actor, invoice)
-          end
-        else
-          {:error, :unauthorized}
-        end
-      end)
-
-    ret
-  end
-
-  defp process_refund_payment(%User{} = actor, %Invoice{} = invoice) do
-    case refund_payment_on_stripe(invoice) do
-      {:ok, %Stripe.Refund{status: "failed"} = refund} ->
-        Logger.error(%{message: "Refund failed", refund: refund})
-        process_refund_updated(refund, invoice.id, actor)
-
-      {:ok, %Stripe.Refund{status: "canceled"} = refund} ->
-        Logger.error(%{message: "Refund canceled", refund: refund})
-        process_refund_updated(refund, invoice.id, actor)
-
-      {:ok, %Stripe.Refund{status: "requires_action"} = refund} ->
-        Logger.info(%{message: "Refund requires action", refund: refund})
-        # This should eventually succeed asynchronously.
-        process_refund_updated(refund, invoice.id, actor)
-
-      {:ok, %Stripe.Refund{status: "succeeded"} = refund} ->
-        process_refund_updated(refund, invoice.id, actor)
-
-      {:ok, %Stripe.Refund{status: "pending"} = refund} ->
-        process_refund_updated(refund, invoice.id, actor)
-
-      {:error, %Stripe.Error{} = err} ->
-        {:error, err}
-    end
-  end
-
-  defp refund_payment_on_stripe(%Invoice{} = invoice) do
-    case stripe_mod().retrieve_session(invoice.stripe_session_id, []) do
-      {:ok, session} ->
-        case stripe_mod().retrieve_payment_intent(session.payment_intent, %{}, []) do
-          {:ok, %Stripe.PaymentIntent{charges: %{data: [%{id: charge_id}]}}} ->
-            # https://stripe.com/docs/connect/destination-charges#issuing-refunds
-            case stripe_mod().create_refund(
-                   %{
-                     charge: charge_id,
-                     reverse_transfer: true,
-                     refund_application_fee: true
-                   },
-                   []
-                 ) do
-              {:ok, refund} ->
-                {:ok, refund}
-
-              {:error, err} ->
-                Logger.error(%{message: "Refund failed.", error: err})
-                {:error, err}
-            end
-
-          {:error, err} ->
-            Logger.error(%{
-              message: "Failed to retrieve payment intent while refunding.",
-              error: err
-            })
-
-            {:error, err}
-        end
-
-      {:error, err} ->
-        Logger.error(%{message: "Failed to retrieve session while refunding.", error: err})
-        {:error, err}
-    end
-  end
-
-  @doc """
-  Webhook handler for when a Stripe refund has been updated.
-  """
-  # credo:disable-for-next-line Credo.Check.Refactor.CyclomaticComplexity
-  def process_refund_updated(%Stripe.Refund{} = refund, invoice_id, actor \\ nil) do
-    {:ok, ret} =
-      Repo.transaction(fn ->
-        assignments =
-          case refund.status do
-            "succeeded" ->
-              [
-                status: :refunded,
-                refund_status: :succeeded,
-                stripe_refund_id: refund.id,
-                refund_failure_reason: nil
-              ]
-
-            "failed" ->
-              [
-                refund_status: :failed,
-                stripe_refund_id: refund.id,
-                refund_failure_reason: refund.failure_reason
-              ]
-
-            "canceled" ->
-              [refund_status: :canceled, stripe_refund_id: refund.id, refund_failure_reason: nil]
-
-            "pending" ->
-              [refund_status: :pending, stripe_refund_id: refund.id, refund_failure_reason: nil]
-
-            "requires_action" ->
-              [
-                refund_status: :requires_action,
-                stripe_refund_id: refund.id,
-                refund_failure_reason: nil
-              ]
-          end
-
-        assignments =
-          if is_nil(actor) do
-            assignments
-          else
-            assignments ++ [refunded_by_id: actor.id]
-          end
-
-        update_res =
-          if is_nil(invoice_id) do
-            from(i in Invoice,
-              where: i.status == :succeeded and i.stripe_refund_id == ^refund.id,
-              select: i.id
-            )
-          else
-            from(i in Invoice, where: i.status == :succeeded and i.id == ^invoice_id, select: i.id)
-          end
-          |> Repo.update_all(set: assignments)
-
-        case update_res do
-          {1, [invoice_id]} ->
-            case from(e in Event,
-                   join: i in assoc(e, :invoice),
-                   where: i.id == ^invoice_id,
-                   select: e,
-                   preload: [
-                     :actor,
-                     :commission,
-                     invoice: [],
-                     attachments: [:upload, :thumbnail, :preview]
-                   ]
-                 )
-                 |> Repo.one() do
-              %Event{} = ev -> {:ok, ev}
-              nil -> {:error, :event_not_found}
-            end
-
-          {0, _} ->
-            {:error, :invoice_not_found}
-        end
-      end)
-
-    case ret do
-      {:ok, event} ->
-        if event.invoice.refund_status == :succeeded do
-          actor =
-            cond do
-              is_nil(actor) && event.invoice.refunded_by_id ->
-                %User{id: event.invoice.refunded_by_id} |> Repo.reload!()
-
-              !is_nil(actor) ->
-                actor
-
-              true ->
-                raise "Bad state: an invoice was succeeded that didn't already have a refunded_by_id."
-            end
-
-          create_event(
-            :refund_processed,
-            actor,
-            event.commission,
-            true,
-            [],
-            %{
-              amount: Money.new(refund.amount, String.to_atom(String.upcase(refund.currency)))
-            }
-          )
-        else
-          Notifications.invoice_refund_updated(event.commission, event, actor)
-        end
-
-        Notifications.commission_event_updated(event.commission, event, actor)
-        {:ok, event.invoice}
-
-      {:error, err} ->
-        Logger.error(%{message: "Failed to update invoice status to refunded", error: err})
-        {:error, :internal_error}
-    end
-  end
-
   ## Deletion
 
   @doc """
@@ -1663,11 +1134,5 @@ defmodule Banchan.Commissions do
       %{event | attachments: new_attachments},
       actor
     )
-  end
-
-  ## Misc utilities
-
-  defp stripe_mod do
-    Application.get_env(:banchan, :stripe_mod)
   end
 end
