+++ lib/banchan_web/controllers/stripe_connect_webhook_controller.ex
@@ -6,7 +6,7 @@ defmodule BanchanWeb.StripeConnectWebhookController do
 
   require Logger
 
-  alias Banchan.Commissions
+  alias Banchan.Payments
   alias Banchan.Studios
 
   def webhook(conn, _params) do
@@ -32,7 +32,7 @@ defmodule BanchanWeb.StripeConnectWebhookController do
 
   defp handle_event(%Stripe.Event{type: "checkout.session.completed"} = event, conn) do
     if event.data.object.payment_status == "paid" do
-      Commissions.process_payment_succeeded!(event.data.object)
+      Payments.process_payment_succeeded!(event.data.object)
     end
 
     conn
@@ -41,7 +41,7 @@ defmodule BanchanWeb.StripeConnectWebhookController do
   end
 
   defp handle_event(%Stripe.Event{type: "checkout.session.expired"} = event, conn) do
-    Commissions.process_payment_expired!(event.data.object)
+    Payments.process_payment_expired!(event.data.object)
 
     conn
     |> resp(200, "OK")
@@ -50,7 +50,7 @@ defmodule BanchanWeb.StripeConnectWebhookController do
 
   defp handle_event(%Stripe.Event{type: "payout." <> type} = event, conn) do
     Logger.debug("Got a new payout event of type payout.#{type}. Updating database.")
-    Studios.process_payout_updated!(event.data.object)
+    Payments.process_payout_updated!(event.data.object)
 
     conn
     |> resp(200, "OK")
@@ -58,7 +58,7 @@ defmodule BanchanWeb.StripeConnectWebhookController do
   end
 
   defp handle_event(%Stripe.Event{type: "charge.refund.updated"} = event, conn) do
-    {:ok, _} = Commissions.process_refund_updated(event.data.object, nil)
+    {:ok, _} = Payments.process_refund_updated(event.data.object, nil)
 
     conn
     |> resp(200, "OK")
